[TASK] databroker dataprocessing in separate module
