+++ src/accml/app/tune/tune_response_analysis.py
@@ -1,44 +1,14 @@
-from collections import defaultdict
-from itertools import zip_longest
-from typing import Dict, Sequence
+from typing import Sequence
 
 import numpy as np
-from numpy.polynomial import Polynomial
+from scipy.linalg import lstsq
+
+from bact_math_utils.linear_fit import x_to_cov, cov_to_std
 
 from .model import (
-    MeasuredTuneResponseItem,
     MeasuredTuneResponsePerPowerConverter,
-    MeasuredTuneResponse,
-    TuneResponse,
-    RandomVariableMomenta,
-    TuneResponseCollection,
+    TuneResponseCollection, MeasuredTuneResponse, RandomVariableMomenta, TuneResponse,
 )
-
-
-def data_to_model(data: Dict[str, Sequence[float]]) -> MeasuredTuneResponse:
-    """
-    For data as stored by bluesky / databroker
-    """
-    # combine data per magnet
-    tmp = defaultdict(list)
-    for dev_name, val, tune_x, tune_y in zip_longest(
-        data["device_name"],
-        data["channel_value"],
-        data["tune-x-sig"],
-        data["tune-y-sig"],
-    ):
-        tmp[str(dev_name)].append(
-            MeasuredTuneResponseItem(value=float(val), x=float(tune_x), y=float(tune_y))
-        )
-
-    # Now put it in a model
-    r = MeasuredTuneResponse(
-        col=[
-            MeasuredTuneResponsePerPowerConverter(pc_name=dev_name, col=items)
-            for dev_name, items in tmp.items()
-        ]
-    )
-    return r
 
 
 def fit_line(indep: Sequence[float], dep: Sequence[float]) -> RandomVariableMomenta:
@@ -49,9 +19,10 @@ def fit_line(indep: Sequence[float], dep: Sequence[float]) -> RandomVariableMome
 
         orbit response profits from leastsquare `lstsg` so perhaps also to use it here
     """
-    coeffs, info = Polynomial.fit(indep, dep, deg=1, full=True)
-    residuals, rank, singular_values, rcond = info
-    return RandomVariableMomenta(mean=float(coeffs.coef[1]), std=float(np.nan))
+    X = np.vstack([np.ones(len(indep)), indep]).T
+    p, res, rnk, s = lstsq(X, dep)
+    dp = cov_to_std(x_to_cov(X, res, N=len(indep), p=2))
+    return RandomVariableMomenta(mean=float(p[1]), std=float(dp[1]))
 
 
 def fit_one_power_converter(data: MeasuredTuneResponsePerPowerConverter):
